Update services and tasks
