+++ src/services/files.py
@@ -14,3 +14,7 @@ class FileService:
         async with async_open(path, "wb") as file:
             await file.write(text.encode())
             file.seek(0)
+
+    @classmethod
+    def delete(cls, path: str) -> None:
+        os.remove(path)
