+++ src/services/paste.py
@@ -1,3 +1,4 @@
+from datetime import datetime, timezone
 from time import time
 
 from fastapi import status
@@ -27,8 +28,7 @@ class PasteService:
         period_name: str | None = None,
     ) -> int:
         paste = Paste()
-        path = f"pastes/{time()*1000}.txt"
-        paste.path = str(path)
+        paste.path = f"pastes/{time()*1000}.txt"
 
         format: Format | None = await FormatRepository.get_one_or_none(
             session, name=format_name
@@ -38,10 +38,14 @@ class PasteService:
             raise HTTPException(status.HTTP_404_NOT_FOUND, "Format not found")
 
         paste.format_id = format.id
+        paste.created_at = datetime.now(timezone.utc)
 
-        session.add(paste)
-        await session.flush([paste])
-
+        try:
+            await FileService.create(str(BASE_DIR / paste.path), text)
+        except Exception:
+            await session.close()
+            raise HTTPException(status.HTTP_400_BAD_REQUEST)
+        
         if period_name is not None:
             period: Period = await PeriodRepository.get_one_or_none(
                 session, name=period_name
@@ -49,20 +53,16 @@ class PasteService:
             if period is None:
                 await session.close()
                 raise HTTPException(status.HTTP_404_NOT_FOUND, "Period not found")
+            
             paste.expire_at = paste.created_at + period.duration
 
-        try:
-            await FileService.create(str(BASE_DIR / path), text)
-        except Exception:
-            await session.close()
-            raise HTTPException(status.HTTP_400_BAD_REQUEST)
-
+        session.add(paste)
         await session.commit()
 
         if paste.expire_at is not None:
             delete_paste.apply_async((paste.id,), eta=paste.expire_at)
 
-        return paste.id
+        return paste
 
     @classmethod
     async def get(cls, session: AsyncSession, id: int) -> str:
@@ -80,3 +80,12 @@ class PasteService:
             "expire_at": paste.expire_at if paste.expire_at is not None else None,
             "format": format.name,
         }
+
+
+@celery_app.task
+def delete_paste(id: int) -> None:
+    with db_helper.sync_session_factory() as session:
+        paste: Paste = PasteSyncRepository.get_by_id(session, id)
+        FileService.delete(str(BASE_DIR / paste.path))
+        session.delete(paste)
+        session.commit()
