+++ src/tasks/tasks.py
@@ -1,16 +1,3 @@
-import os
+from src.services.paste import delete_paste
 
-from .celery import celery_app
-from src.core.db_helper import db_helper
-from src.core.models import Paste
-from src.core.repositories import PasteSyncRepository
-from src.core.config import BASE_DIR
-
-
-@celery_app.task
-def delete_paste(id: int) -> None:
-    with db_helper.sync_session_factory() as session:
-        paste: Paste = PasteSyncRepository.get_by_id(session, id)
-        os.remove(str(BASE_DIR / paste.path))
-        session.delete(paste)
-        session.commit()
+__all__ = ("delete_paste",)
